Organize project as packages
