+++ src/main/java/cattis/command/ExitCommand.java
@@ -1,4 +1,15 @@
 package cattis.command;
 
-public class ExitCommand {
+import cattis.Cattis;
+import cattis.exception.CattisException;
+
+public class ExitCommand extends Command {
+    @Override
+    public void execute(Cattis cattis) throws CattisException {
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
 }
