+++ src/main/java/cattis/command/MarkCommand.java
@@ -1,4 +1,19 @@
 package cattis.command;
 
-public class MarkCommand {
+import cattis.Cattis;
+import cattis.Constants;
+import cattis.exception.CattisException;
+
+public class MarkCommand extends Command {
+    private final int taskIndex;
+
+    public MarkCommand(int taskIndex) {
+        this.taskIndex = taskIndex;
+    }
+
+    @Override
+    public void execute(Cattis cattis) throws CattisException {
+        cattis.getTaskList().mark(taskIndex);
+        System.out.printf((Constants.MARK_TASK_MSG), cattis.getTaskList().get(taskIndex));
+    }
 }
