+++ src/main/java/cattis/command/UnmarkCommand.java
@@ -1,4 +1,19 @@
 package cattis.command;
 
-public class UnmarkCommand {
+import cattis.Cattis;
+import cattis.Constants;
+import cattis.exception.CattisException;
+
+public class UnmarkCommand extends Command {
+    private final int taskIndex;
+
+    public UnmarkCommand(int taskIndex) {
+        this.taskIndex = taskIndex;
+    }
+
+    @Override
+    public void execute(Cattis cattis) throws CattisException {
+        cattis.getTaskList().unmark(taskIndex);
+        System.out.printf((Constants.UNMARK_TASK_MSG), cattis.getTaskList().get(taskIndex));
+    }
 }
