+++ src/main/java/cattis/task/Task.java
@@ -1,6 +1,6 @@
-package cattis.Task;
+package cattis.task;
 
-import cattis.Exceptions.CattisException;
+import cattis.exception.CattisException;
 
 import java.util.Arrays;
 import java.util.List;
